+++ sdk/@launchdarkly/observability-android/lib/src/main/kotlin/com/launchdarkly/observability/plugin/Observability.kt
@@ -73,7 +73,6 @@ class Observability(
     override fun register(client: LDClient, metadata: EnvironmentMetadata?) {
         this.client = client
         val sdkKey = metadata?.credential ?: ""
-
         if (mobileKey == sdkKey) {
             LDObserve.context = ObservabilityContext(
                 sdkKey = sdkKey,
@@ -87,9 +86,14 @@ class Observability(
     }
 
     override fun getHooks(metadata: EnvironmentMetadata?): MutableList<Hook> {
-        return Collections.singletonList(
-            ObservabilityHook(withSpans = true, withValue = true) { observabilityClient?.getTracer() }
+        val exporter = ObservabilityHookExporter(
+            withSpans = true,
+            withValue = true,
+            tracerProvider = { observabilityClient?.getTracer() },
+            contextFriendlyName = options.contextFriendlyName
         )
+        LDObserve.hookProxy = ObservabilityHookProxy(exporter)
+        return Collections.singletonList(ObservabilityHook(exporter))
     }
 
     override fun onPluginsReady(result: RegistrationCompleteResult?, metadata: EnvironmentMetadata?) {