+++ sdk/@launchdarkly/observability-android/lib/src/main/kotlin/com/launchdarkly/observability/replay/capture/ExportDiffManager.kt
@@ -13,10 +13,14 @@ class ExportDiffManager(
     private val currentImages = mutableListOf<ExportFrame.RemoveImage>()
     private val currentImagesIndex = mutableMapOf<ImageSignature, Int>()
     private val lock = Any()
-    private val format = ExportFrame.ExportFormat.Webp(quality = 30)
+    private val format = ExportFrame.ExportFormat.Webp(quality = 30) //ExportFrame.ExportFormat.Jpeg(quality = 0.3f)
+
     private var keyFrameId = 0
 
-    fun createCaptureEvent(rawFrame: ImageCaptureService.RawFrame, session: String): ExportFrame? {
+    fun createCaptureEvent(
+        rawFrame: ImageCaptureService.RawFrame,
+        session: String
+    ): ExportFrame? {
         synchronized(lock) {
             val tiledFrame = tileDiffManager.computeTiledFrame(rawFrame) ?: return null
             return createCaptureEventInternal(tiledFrame, session)