+++ sdk/@launchdarkly/observability-android/CHANGELOG.md
@@ -1,5 +1,17 @@
 # Changelog
 
+## [0.29.0](https://github.com/launchdarkly/observability-sdk/compare/launchdarkly-observability-android-0.28.0...launchdarkly-observability-android-0.29.0) (2026-03-05)
+
+
+### Features
+
+* Android Observability hook  proxy for MAUI ([#409](https://github.com/launchdarkly/observability-sdk/issues/409)) ([8d610bd](https://github.com/launchdarkly/observability-sdk/commit/8d610bdd700fd2878e58095e5fd8b98ef2765df6))
+
+
+### Bug Fixes
+
+* reset nodeIds during fullsnapshot ([#412](https://github.com/launchdarkly/observability-sdk/issues/412)) ([d1eb13d](https://github.com/launchdarkly/observability-sdk/commit/d1eb13db8907f17406f45a2a7e3b76aec3a7c1ec))
+
 ## [0.28.0](https://github.com/launchdarkly/observability-sdk/compare/launchdarkly-observability-android-0.27.0...launchdarkly-observability-android-0.28.0) (2026-03-02)